ecr metadata service
