+++ containers/ecr-viewer/src/app/view-data/services/EcrMetadataService.ts
@@ -1,0 +1,98 @@
+import { formatDateTime } from "@/app/format-service";
+import {
+  DisplayDataProps,
+  PathMappings,
+  evaluateData,
+  extractFacilityAddress,
+} from "@/app/utils";
+import { Bundle } from "fhir/r4";
+import { evaluate } from "fhirpath";
+
+export interface ReportableConditions {
+  [condition: string]: {
+    [trigger: string]: Set<string>;
+  };
+}
+
+/**
+ * Evaluates eCR metadata from the FHIR bundle and formats it into structured data for display.
+ * @param fhirBundle - The FHIR bundle containing eCR metadata.
+ * @param mappings - The object containing the fhir paths.
+ * @returns An object containing evaluated and formatted eCR metadata.
+ */
+export const evaluateEcrMetadata = (
+  fhirBundle: Bundle,
+  mappings: PathMappings,
+) => {
+  const rrDetails = evaluate(fhirBundle, mappings.rrDetails);
+
+  let reportableConditionsList: ReportableConditions = {};
+
+  for (const condition of rrDetails) {
+    let name = condition.valueCodeableConcept.coding[0].display;
+    const triggers = condition.extension
+      .filter(
+        (x: { url: string; valueString: string }) =>
+          x.url ===
+          "http://hl7.org/fhir/us/ecr/StructureDefinition/us-ph-determination-of-reportability-rule-extension",
+      )
+      .map((x: { url: string; valueString: string }) => x.valueString);
+    if (!reportableConditionsList[name]) {
+      reportableConditionsList[name] = {};
+    }
+
+    for (let i in triggers) {
+      if (!reportableConditionsList[name][triggers[i]]) {
+        reportableConditionsList[name][triggers[i]] = new Set();
+      }
+      condition.performer
+        .map((x: { display: string }) => x.display)
+        .forEach((x: string) =>
+          reportableConditionsList[name][triggers[i]].add(x),
+        );
+    }
+  }
+
+  const eicrDetails: DisplayDataProps[] = [
+    {
+      title: "eICR Identifier",
+      toolTip:
+        "Unique document ID for the eICR that originates from the medical record. Different from the Document ID that NBS creates for all incoming records.",
+      value: evaluate(fhirBundle, mappings.eicrIdentifier)[0],
+    },
+  ];
+  const ecrSenderDetails: DisplayDataProps[] = [
+    {
+      title: "Date/Time eCR Created",
+      value: formatDateTime(
+        evaluate(fhirBundle, mappings.dateTimeEcrCreated)[0],
+      ),
+    },
+    {
+      title: "Sender Software",
+      toolTip: "EHR system used by the sending provider.",
+      value: evaluate(fhirBundle, mappings.senderSoftware)[0],
+    },
+    {
+      title: "Sender Facility Name",
+      value: evaluate(fhirBundle, mappings.senderFacilityName)[0],
+    },
+    {
+      title: "Facility Address",
+      value: extractFacilityAddress(fhirBundle, mappings),
+    },
+    {
+      title: "Facility Contact",
+      value: evaluate(fhirBundle, mappings.facilityContact)[0],
+    },
+    {
+      title: "Facility ID",
+      value: evaluate(fhirBundle, mappings.facilityID)[0],
+    },
+  ];
+  return {
+    eicrDetails: evaluateData(eicrDetails),
+    ecrSenderDetails: evaluateData(ecrSenderDetails),
+    rrDetails: reportableConditionsList,
+  };
+};
